+++ packages/dashboard/src/components/rmf-app/index.tsx
@@ -1,2 +1,3 @@
 export * from './contexts';
 export * from './rmf-app';
+export * from './utils';
