+++ packages/dashboard/src/components/schedule-visualizer/tests/waypoints-overlay.test.tsx
@@ -1,7 +1,8 @@
+import { render, waitFor } from '@testing-library/react';
 import L from 'leaflet';
 import React from 'react';
-import { render, waitFor } from '@testing-library/react';
 import { Map as LMap } from 'react-leaflet';
+import { getPlaces, PlacesContext } from '../../rmf-app';
 import WaypointsOverlay from '../waypoints-overlay';
 import getBuildingMap from './building-map';
 
@@ -9,18 +10,20 @@ test('Render waypoints correctly', async () => {
   const bounds = new L.LatLngBounds([0, 25.7], [-14, 0]);
   const buildingMap = await getBuildingMap();
   const currentLevel = buildingMap.levels[0];
-  const nav_graphs = buildingMap.levels.flatMap((x) => x.nav_graphs);
-  const waypoints = nav_graphs[0].vertices;
+  const places = getPlaces(buildingMap);
+  const waypoints = Object.values(places);
 
   const root = render(
-    <LMap
-      bounds={[
-        [0, 0],
-        [1, 1],
-      ]}
-    >
-      <WaypointsOverlay bounds={bounds} currentLevel={currentLevel} />
-    </LMap>,
+    <PlacesContext.Provider value={places}>
+      <LMap
+        bounds={[
+          [0, 0],
+          [1, 1],
+        ]}
+      >
+        <WaypointsOverlay bounds={bounds} currentLevel={currentLevel} />
+      </LMap>
+    </PlacesContext.Provider>,
   );
 
   await waitFor(() => {
